style: rename package
