+++ src/main/java/rentcompany/RentCompany.java
@@ -1,4 +1,4 @@
-package RentCompany;
+package rentcompany;
 
 import java.util.ArrayList;
 import java.util.List;
