+++ src/test/java/rentcompany/RentCompanyTest.java
@@ -1,4 +1,4 @@
-package RentCompany;
+package rentcompany;
 
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
